рефакторинг: переименование groupChats в wsChat, который включает подключение ко всем чатам в зависимости от chatId и типа чата;
